+++ aioxmpp/presence/__init__.py
@@ -1,0 +1,8 @@
+"""
+:mod:`~aioxmpp.presence` --- Peer presence bookkeeping
+######################################################
+
+.. autoclass:: Service
+"""
+
+from .service import Service